Add Item's methods
